==> code/analysis/00_preprocessing/00_clean-covariates.qmd <==
---
title: "00-00. Tidy covariate data for downstream analyses"
author: "Rachel Smith"
format: html
editor: visual
---

This script loads covariate data from multiple sources and combines it for downstream analyses.

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
base_dir <- "~/Documents/PhD/projects/sgacc_wgcna_grcca/"
objects_dir <- paste0(base_dir, "outputs/objects/")
tables_dir <- paste0(base_dir, "outputs/tables/")
```

## Libraries

```{r libraries, message=FALSE, warning=FALSE}
library(tidyverse)
library(readxl)
library(janitor)
```

## Data

Complete covariates data (from Ajeet)

```{r load_data}
## Complete covariate data
df_complete_covariates <- read_xlsx(paste0(base_dir, "data/14Nov2023_all_covariates.xlsx"))

## Other covariates from Nirmala (includes additional technical covariates, like sgACC RIN)
df_nirmala_covariates <- read.table(
    paste0(
        base_dir, 
        "data/185samples_multinomialDx_46covariates_10evecs.txt"), 
    sep = "\t", header = TRUE) %>% 
    dplyr::select(-contains("evec")) %>% 
    as_tibble() %>% 
    mutate_if(is.character, ~ ifelse(.x == "", NA_character_, .x)) %>% 
    clean_names %>% 
    dplyr::rename("rna_extraction_batch" = "rn_aextraction_batch", 
                  "gc_percent" = "g_cpercent",
                  "dx" = "multi_nomial_dx") %>% 
    dplyr::select(-contains("_1")) %>% # remove duplicate columns
    mutate(sample = tolower(sample))
```

## Define covariate types

```{r define_covariates}
technical_covariates <- c("mapped_percent", "gc_percent", "five_prime_three_prime_bias", 
                          "rin_acsg", "rna_extraction_batch",
                          "library_batch", "pmi", "pmi_confidence", 
                          "source", "ph" # "max_rin", "max_rine"
                           )
biological_covariates <- c("age_death", "sex_at_birth", "race", "bmi", "height", "weight",
                           "marital_status", "manner_death", "education", 
                           "suicide", "brain_weight")
drug_covariates <- c("smoker", "nicotine_cotinine", "alcohol", "opioids",
                      "beta_blockers", "major_stimulants_cocaine_included", "cannabinoids",
                     "minor_stimulants", "anticholinergics", "antidepressants",
                     "anti_epileptics", "anti_histamines", "antipsychotics", 
                     "mood_stabilizers", "sedative_hypnotic_anxiolitics", "benzos",
                     "other_psychotropic_drug", "non_psychiatric")
```

## Format data

Remove covariates that are not included in downstream analyses; format covariate names

```{r format_data}
df_complete_covariates2 <- df_complete_covariates %>% 
    clean_names %>%
    
    # remove some covariates that are repetitive
    dplyr::select(-c(brain_number_5, 
                     verified, 
                     inclusion, 
                     avaialble_short_read_rna_seq, 
                     pseudo_guid, 
                     primary_dx,
                     gender_at_death,
                     sexual_orientation,
                     ethnicity,
                     contains("axi"),
                     cause_death,
                     source,
                     final_neuropath_diagnosis,
                     inhalants_zinhal, # (only 1 non-NA value)
                     hallucinogens_zhall # (everyone is negative)
    )
    ) %>% 
    
    # clean up covariate names
    dplyr::rename("brain_id" = "brain_number_1", 
                  "ph" = "p_h",
                  "max_rine" = "max_ri_ne") %>% 
    dplyr::rename_all(~str_remove(.x, "_z.*")) %>% 
    mutate(sample = paste0(brain_id, sep = "_", tolower(dx)), .before = 1) %>% 
    
    # clean up covariate content
    mutate(suicide = ifelse(manner_death == "Suicide", 1, 0), .before = brain_weight) %>% 
    mutate(dx = case_when(
        dx == "Schizo" ~ "SCZ",
        dx == "Bipolar" ~ "BD",
        TRUE ~ dx
    ),
    marital_status = ifelse(marital_status %in% c("Unknown", "not filled in"), 
                            NA_character_, 
                            marital_status),
    education = ifelse(education %in% c("Unknown", "not filled in"), 
                       NA_character_, 
                       education)
    #manner_death = ifelse(manner_death == "Undetermined", NA_character_, manner_death)
    )

## Check what variables are included in both covariate tables
shared_variables <- intersect(colnames(df_complete_covariates2), 
                              colnames(df_nirmala_covariates))
colnames(df_nirmala_covariates)[!(colnames(df_nirmala_covariates) %in% shared_variables)]
```

Add technical covariates from Nirmala's df; arrange by covariate type (for export)

```{r merge_covariates}
df_covariates <- df_complete_covariates2 %>% 
  left_join(
    df_nirmala_covariates %>% 
      dplyr::select(sample, 
                    mapped_percent, 
                    rin_acsg,
                    source,
                    five_prime_three_prime_bias, 
                    rna_extraction_batch, 
                    library_batch, 
                    gc_percent,
                    benzos),
    by = join_by(sample)
  ) %>% 
    
    dplyr::select(sample, dx, 
                  all_of(c(biological_covariates, technical_covariates, drug_covariates))
    )
```

Convert each covariate to numeric

```{r numeric_covariates}
df_covariates_numeric <- df_covariates %>% 
  dplyr::select(-sample) %>% 
  mutate_if(is.character, ~ ifelse(.x == "NULL", NA_character_, .x)) %>% 
  mutate_if(is.character, ~ as.factor(.x) %>% as.numeric - 1) %>% 
  mutate(sample = df_covariates$sample)
```

## Save

```{r save}
# save(df_covariates, df_covariates_numeric,
#      file = paste0(objects_dir, "22Jan2025_covariates.Rdata")
# )
```

## Export covariates table for supplement

```{r tableS1}
write_xlsx(df_covariates %>% 
               mutate(dx = factor(dx, levels = names(dx_colors))) %>% 
               arrange(dx, sex_at_birth, race, age_death) %>% 
               mutate_at(vars(c(bmi, height, weight)), ~ as.numeric(.x)) %>% 
               dplyr::select(-suicide, -sample), 
           path = paste0(tables_dir, "TableS1.covariate_information.xlsx")
)
```
